+++ src/app/app.component.spec.ts
@@ -2,7 +2,7 @@ import { TestBed, async } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { AppComponent } from './app.component';
 import { TranslateModule } from '@ngx-translate/core';
-import { ElectronService } from './services/electron.service';
+import { ElectronService } from './core/services/electron.service';
 
 describe('AppComponent', () => {
   beforeEach(async(() => {
