+++ src/app/core/services/electron.service.ts
