+++ src/app/home/home.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+
+import { HomeComponent } from './home.component';
+import {HomeRoutingModule} from './home-routing.module';
+import {SharedModule} from '../shared/shared.module';
+
+@NgModule({
+  imports: [
+    HomeRoutingModule,
+    SharedModule
+  ],
+  declarations: [
+    HomeComponent
+  ],
+})
+export class HomeModule {}
